修复OperateRepairActivity中显示新设备的bug

+++ app/src/main/java/com/tianyigps/xiepeng/manager/DatabaseManager.java
@@ -24,13 +24,18 @@ public class DatabaseManager {
     }
 
     //  增，维修
-    public void addRepair(String tNo, String position, String positionPic, String installPic, String explain) {
+    public void addRepair(String tNo, String position, String positionPic, String installPic, String explain, String newtNo) {
+        if (repairExist(tNo)) {
+            modifyRepair(tNo, position, positionPic, installPic, explain, newtNo);
+            return;
+        }
         ContentValues contentValues = new ContentValues();
         contentValues.put("tNo", tNo);
         contentValues.put("position", position);
         contentValues.put("positionPic", positionPic);
         contentValues.put("installPic", installPic);
         contentValues.put("explain", explain);
+        contentValues.put("newtNo", newtNo);
 
         mSqLiteDatabase.beginTransaction();
         try {
@@ -43,8 +48,69 @@ public class DatabaseManager {
         }
     }
 
+    //  增，维修
+    public void addRepair(String tNo, String position, String explain) {
+        addRepairPosition(tNo, position);
+        addRepairExplain(tNo, explain);
+    }
+
+    //  增，维修
+    public void addRepair(String tNo, String position, String explain, String newtNo) {
+        addRepairPosition(tNo, position);
+        addRepairExplain(tNo, explain);
+        addRepairNewtNo(tNo, newtNo);
+    }
+
+    //  增，维修，重载
+    public void addRepairPosition(String tNo, String position) {
+        if (repairExist(tNo)) {
+            modifyRepairPosition(tNo, position);
+            return;
+        }
+        addRepair(tNo, position, null, null, null, null);
+    }
+
+    //  增，维修，重载
+    public void addRepairPositionPic(String tNo, String positionPic) {
+        if (repairExist(tNo)) {
+            modifyRepairPositionPic(tNo, positionPic);
+            return;
+        }
+        addRepair(tNo, null, positionPic, null, null, null);
+    }
+
+    //  增，维修，重载
+    public void addRepairInstallPic(String tNo, String installPic) {
+        if (repairExist(tNo)) {
+            modifyRepairInstallPic(tNo, installPic);
+            return;
+        }
+        addRepair(tNo, null, null, installPic, null, null);
+    }
+
+    //  增，维修，重载
+    public void addRepairExplain(String tNo, String explain) {
+        if (repairExist(tNo)) {
+            modifyRepairExplain(tNo, explain);
+            return;
+        }
+        addRepair(tNo, null, null, null, explain, null);
+    }
+
+    //  增，维修，重载
+    public void addRepairNewtNo(String tNo, String newtNo) {
+        if (repairExist(tNo)) {
+            modifyRepairnewtNo(tNo, newtNo);
+            return;
+        }
+        addRepair(tNo, null, null, null, null, newtNo);
+    }
+
     //  删，维修
     public void deleteRepair(String tNo) {
+        if (!repairExist(tNo)) {
+            return;
+        }
         mSqLiteDatabase.beginTransaction();
         try {
             mSqLiteDatabase.delete(Data.DATA_TAB_REPAIR
@@ -59,13 +125,125 @@ public class DatabaseManager {
     }
 
     //  改，维修
-    public void modifyRepair(String tNo, String position, String positionPic, String installPic, String explain) {
+    public void modifyRepair(String tNo, String position, String positionPic, String installPic, String explain, String newtNo) {
+        if (!repairExist(tNo)) {
+            return;
+        }
         ContentValues contentValues = new ContentValues();
-        contentValues.put("tNo", tNo);
         contentValues.put("position", position);
         contentValues.put("positionPic", positionPic);
         contentValues.put("installPic", installPic);
         contentValues.put("explain", explain);
+        contentValues.put("newtNo", newtNo);
+
+        mSqLiteDatabase.beginTransaction();
+        try {
+            mSqLiteDatabase.update(Data.DATA_TAB_REPAIR
+                    , contentValues
+                    , "tNo=?"
+                    , new String[]{tNo});
+            mSqLiteDatabase.setTransactionSuccessful();
+        } catch (Exception e) {
+            Log.e(TAG, e + "SqLiteDatabase update error");
+        } finally {
+            mSqLiteDatabase.endTransaction();
+        }
+    }
+
+    //  改，维修，重载
+    public void modifyRepairPosition(String tNo, String position) {
+        if (!repairExist(tNo)) {
+            return;
+        }
+        ContentValues contentValues = new ContentValues();
+        contentValues.put("position", position);
+
+        mSqLiteDatabase.beginTransaction();
+        try {
+            mSqLiteDatabase.update(Data.DATA_TAB_REPAIR
+                    , contentValues
+                    , "tNo=?"
+                    , new String[]{tNo});
+            mSqLiteDatabase.setTransactionSuccessful();
+        } catch (Exception e) {
+            Log.e(TAG, e + "SqLiteDatabase update error");
+        } finally {
+            mSqLiteDatabase.endTransaction();
+        }
+    }
+
+    //  改，维修，重载
+    public void modifyRepairPositionPic(String tNo, String positionPic) {
+        if (!repairExist(tNo)) {
+            return;
+        }
+        ContentValues contentValues = new ContentValues();
+        contentValues.put("positionPic", positionPic);
+
+        mSqLiteDatabase.beginTransaction();
+        try {
+            mSqLiteDatabase.update(Data.DATA_TAB_REPAIR
+                    , contentValues
+                    , "tNo=?"
+                    , new String[]{tNo});
+            mSqLiteDatabase.setTransactionSuccessful();
+        } catch (Exception e) {
+            Log.e(TAG, e + "SqLiteDatabase update error");
+        } finally {
+            mSqLiteDatabase.endTransaction();
+        }
+    }
+
+    //  改，维修，重载
+    public void modifyRepairInstallPic(String tNo, String installPic) {
+        if (!repairExist(tNo)) {
+            return;
+        }
+        ContentValues contentValues = new ContentValues();
+        contentValues.put("installPic", installPic);
+
+        mSqLiteDatabase.beginTransaction();
+        try {
+            mSqLiteDatabase.update(Data.DATA_TAB_REPAIR
+                    , contentValues
+                    , "tNo=?"
+                    , new String[]{tNo});
+            mSqLiteDatabase.setTransactionSuccessful();
+        } catch (Exception e) {
+            Log.e(TAG, e + "SqLiteDatabase update error");
+        } finally {
+            mSqLiteDatabase.endTransaction();
+        }
+    }
+
+    //  改，维修，重载
+    public void modifyRepairExplain(String tNo, String explain) {
+        if (!repairExist(tNo)) {
+            return;
+        }
+        ContentValues contentValues = new ContentValues();
+        contentValues.put("explain", explain);
+
+        mSqLiteDatabase.beginTransaction();
+        try {
+            mSqLiteDatabase.update(Data.DATA_TAB_REPAIR
+                    , contentValues
+                    , "tNo=?"
+                    , new String[]{tNo});
+            mSqLiteDatabase.setTransactionSuccessful();
+        } catch (Exception e) {
+            Log.e(TAG, e + "SqLiteDatabase update error");
+        } finally {
+            mSqLiteDatabase.endTransaction();
+        }
+    }
+    //  改，维修，重载
+    public void modifyRepairnewtNo(String tNo, String newtNo) {
+        if (!repairExist(tNo)) {
+            return;
+        }
+        ContentValues contentValues = new ContentValues();
+        contentValues.put("newtNo", newtNo);
 
         mSqLiteDatabase.beginTransaction();
         try {
@@ -82,11 +260,11 @@ public class DatabaseManager {
     }
 
     //  查，维修
-    public Cursor getRepair() {
+    public Cursor getRepairs() {
         mSqLiteDatabase.beginTransaction();
         try {
             Cursor cursor = mSqLiteDatabase.query(Data.DATA_TAB_REPAIR
-                    , new String[]{"tNo, position, positionPic, installPic, explain"}
+                    , new String[]{"tNo, position, positionPic, installPic, explain, newtNo"}
                     , null
                     , null
                     , null, null, null);
@@ -106,7 +284,7 @@ public class DatabaseManager {
         mSqLiteDatabase.beginTransaction();
         try {
             Cursor cursor = mSqLiteDatabase.query(Data.DATA_TAB_REPAIR
-                    , new String[]{"tNo, position, positionPic, installPic, explain"}
+                    , new String[]{"tNo, position, positionPic, installPic, explain, newtNo"}
                     , "tNo=?"
                     , new String[]{tNo}
                     , null, null, null);
@@ -119,6 +297,12 @@ public class DatabaseManager {
             mSqLiteDatabase.endTransaction();
         }
         return null;
+    }
+
+    //  查，是否存在该order
+    public boolean repairExist(String tNo) {
+        Cursor cursor = getRepair(tNo);
+        return cursor.moveToFirst();
     }
 
     //=============================================华丽的分割线===================================================
